Create MAD database reference file
If madDB is set in config then a JSON will be created when the bot starts that includes basic MAD info about instances/devices/areas.  This info will be used for future features to help decrease the number of connections to the database. Bot will need to be restarted if any changes are made to MAD.